Patch fix for new restricted freqs
Changing the checks to look at a float range since the new freqs need
better precision.

+++ utils/experiment_options/experimentoptions.py
@@ -144,7 +144,7 @@ class ExperimentOptions:
         self._analog_rx_attenuator = params[9]  # dB
         self._tdiff = params[10] # ns
         self._phase_sign = params[11]
-        self._intf_offset = [float(params[12]), float(params[13]), float(params[14])]  
+        self._intf_offset = [float(params[12]), float(params[13]), float(params[14])]
         # interferometer offset from
         # midpoint of main, metres [x, y, z] where x is along line of antennas, y is along array
         # normal and z is altitude difference, in m.
@@ -169,7 +169,7 @@ class ExperimentOptions:
             splitup = line.split("=")
             if len(splitup) == 2:
                 if splitup[0] == 'default' or splitup[0] == 'default ':
-                    self.__default_freq = int(splitup[1])  # kHz
+                    self.__default_freq = float(splitup[1])  # kHz
                     restricted.remove(line)
                     break
         else: #no break
@@ -181,7 +181,7 @@ class ExperimentOptions:
             if len(splitup) != 2:
                 raise Exception('Problem with Restricted Frequency: A Range Len != 2')
             try:
-                splitup = [int(float(freq)) for freq in splitup]  # convert to ints
+                splitup = [float(freq) for freq in splitup]  # convert to ints
             except ValueError:
                 raise ValueError('Error parsing Restrict.Dat Frequency Ranges, Invalid Literal')
             restricted_range = tuple(splitup)
@@ -286,7 +286,7 @@ class ExperimentOptions:
     @property
     def max_number_of_filter_taps_per_stage(self):
         return self._max_number_of_filter_taps_per_stage
-  
+
     @property
     def site_id(self):
         return self._site_id
